Move the gRPC plugin into its own directory. (#47)
