+++ opencensus/plugins/grpc/internal/client_filter.cc
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "opencensus/plugins/internal/client_filter.h"
+#include "opencensus/plugins/grpc/internal/client_filter.h"
 
 #include "absl/strings/str_cat.h"
 #include "absl/strings/string_view.h"
-#include "opencensus/plugins/grpc_plugin.h"
-#include "opencensus/plugins/internal/measures.h"
+#include "opencensus/plugins/grpc/grpc_plugin.h"
+#include "opencensus/plugins/grpc/internal/measures.h"
 #include "opencensus/stats/stats.h"
 #include "src/core/lib/surface/call.h"
