+++ opencensus/plugins/grpc/internal/grpc_plugin.cc
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "opencensus/plugins/grpc_plugin.h"
-#include "opencensus/plugins/internal/channel_filter.h"
-#include "opencensus/plugins/internal/client_filter.h"
-#include "opencensus/plugins/internal/measures.h"
-#include "opencensus/plugins/internal/server_filter.h"
+#include "opencensus/plugins/grpc/grpc_plugin.h"
+#include "opencensus/plugins/grpc/internal/channel_filter.h"
+#include "opencensus/plugins/grpc/internal/client_filter.h"
+#include "opencensus/plugins/grpc/internal/measures.h"
+#include "opencensus/plugins/grpc/internal/server_filter.h"
 #include "src/core/lib/surface/channel_init.h"
 
 namespace opencensus {
